+++ packages/headless/src/api/analytics/coveo.analytics/history-store.test.ts
@@ -12,9 +12,9 @@ import {
   MAX_NUMBER_OF_HISTORY_ELEMENTS,
   MAX_VALUE_SIZE,
   STORE_KEY,
-} from './history-store.js';
+} from 'coveo.analytics/dist/esm/history.mjs';
+import type {WebStorage} from 'coveo.analytics/dist/esm/storage.mjs';
 import HistoryStore from './history-store.js';
-import type {WebStorage} from './storage.js';
 
 describe('HistoryStore', () => {
   let mockStorage: Mocked<WebStorage>;